engine: gate lookup_with_hint on per-table x-sortedness
The hint uniqueness argument in lookup_with_hint (GH #602) requires
the table's x-values to be non-decreasing: only then does a hint h
satisfying table[h-1].0 < index <= table[h].0 uniquely identify the
binary-search lower bound. No import path (XMILE, protobuf/DB, MDL,
or variable.rs parse_table) validates or sorts x-values, so an
unsorted table reaching lookup_with_hint could accept a stale hint,
run lookup_tail on the wrong low, and return a completely wrong result
-- diverging silently from lookup (the wasm oracle).

Review found the reviewer's minimal reproducer:
  t = [(0,0), (10,1000), (5,50), (20,200)]
  first call at 0.25 seeds hint=1; second call at 7.0 accepts hint=1
  (table[0].0=0 < 7 <= table[1].0=10), runs lookup_tail at low=1,
  and returns 700 instead of the correct 70.

Fix: at Vm::new, classify each table with table_x_is_sorted
(windows(2).all(|w| w[0].0 <= w[1].0)); unsorted tables (including
any with NaN x-values, which compare false) are assigned
LOOKUP_HINT_DISABLED (u32::MAX) in lookup_memo. lookup_with_hint
checks the sentinel first: if disabled, it delegates to lookup verbatim
and returns without touching the hint slot, making it structurally
impossible for the miss path to overwrite the sentinel. Sorted tables
continue to use 0 (never a valid hint, seeds on first interior lookup).

No import-boundary validation or sorting is added (separate issue per
manager decision: that changes behavior of DB-stored models).

The lookup_hint_tests module is split to vm_lookup_hint_tests.rs to
keep vm.rs under the 6000-line per-file limit.

Tests added: the reviewer's exact reproducer (bit-exact with lookup,
sentinel never overwritten across multiple calls), a proptest over
unsorted tables with the forced sentinel (bit-exact + sentinel stable),
table_x_is_sorted unit tests (sorted/unsorted/NaN/edge cases), and a
Vm::new integration test (sorted GF round-trips correctly).
